Return the correct Variant ID in case the Chain ID is not in the ChainIdLookup
This PR addresses an issue encountered while using this library with the WAX testnet. Specifically, when a new request was generated and the VariantId function was called, it couldn't locate the chain alias as the ChainIdLookup dictionary did not contain the specific chain id. This would subsequently lead to a return of a KeyValuePair containing chain_alias paired with the provided chain_id. This mismatch became problematic during parsing since the chain_id is expected to be a string, whereas the chain_alias is long.

Upon consulting the Greymass documentation, it's evident that if a chain_alias is not identified, we should return the chain_id in its stead. This PR ensures that behavior is implemented correctly.

Relevant documentation guiding this change can be found here

https://github.com/greymass/eosio-signing-request/blob/master/protocol-specification.md